refactor(git-utils): final clean up of types
