--- packages/amplication-git-utils/src/git/index.ts
@@ -1,1 +1,0 @@
-export * from "./dto";
